Refactor the DB code.
